services/SafeZoneAnalyticsAPI/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-json-logger==3.2.1
redis==6.2.0
